move tuning menu stuff in separate device
