+++ Kbuild
@@ -1,2 +1,2 @@
 obj-m := hid-fanatec.o 
-hid-fanatec-y := hid-ftec.o hid-ftecff.o
+hid-fanatec-y := hid-ftec.o hid-ftecff.o hid-ftecff-tuning.o
