+++ hid-ftec.h
@@ -78,6 +78,12 @@ struct ftecff_slot {
 	u8 cmd;
 };
 
+struct ftec_tuning_classdev {
+	struct device *dev;
+	// the data from the last update we got from the device, shifted by 1
+	u8 ftec_tuning_data[FTEC_TUNING_REPORT_SIZE];
+};
+
 struct ftec_drv_data {
 	unsigned long quirks;
 	spinlock_t report_lock; /* Protect output HID report */
@@ -95,9 +101,33 @@ struct ftec_drv_data {
 	u16 led_state;
 	struct led_classdev *led[LEDS];
 #endif    
-	// the data from the last update we got from the device, shifted by 1
-	u8 ftec_tuning_data[FTEC_TUNING_REPORT_SIZE];
 	u8 wheel_id;
+	struct ftec_tuning_classdev tuning;
+};
+
+#define FTEC_TUNING_ATTRS \
+	FTEC_TUNING_ATTR(SLOT, 0x02, "Slot", ftec_conv_noop_to, ftec_conv_noop_from, 1, 5) \
+	FTEC_TUNING_ATTR(SEN, 0x03, "Sensivity", ftec_conv_sens_to, ftec_conv_sens_from, 90, 0) \
+	FTEC_TUNING_ATTR(FF, 0x04, "Force Feedback Strength", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
+	FTEC_TUNING_ATTR(SHO, 0x05, "Wheel Vibration Motor", ftec_conv_times_ten, ftec_conv_div_ten, 0, 100) \
+	FTEC_TUNING_ATTR(BLI, 0x06, "Break Level Indicator", ftec_conv_noop_to, ftec_conv_noop_from, 0, 101) \
+	FTEC_TUNING_ATTR(DRI, 0x09, "Drift Mode", ftec_conv_signed_to, ftec_conv_noop_from, -5, 3) \
+	FTEC_TUNING_ATTR(FOR, 0x0a, "Force Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
+	FTEC_TUNING_ATTR(SPR, 0x0b, "Spring Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
+	FTEC_TUNING_ATTR(DPR, 0x0c, "Damper Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
+	FTEC_TUNING_ATTR(NDP, 0x0d, "Natural Damber", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
+	FTEC_TUNING_ATTR(NFR, 0x0e, "Natural Friction", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
+	FTEC_TUNING_ATTR(FEI, 0x11, "Force Effect Intensity", ftec_conv_noop_to, ftec_conv_steps_ten, 0, 100) \
+	FTEC_TUNING_ATTR(INT, 0x14, "FFB Interpolation Filter", ftec_conv_noop_to, ftec_conv_noop_from, 0, 20) \
+	FTEC_TUNING_ATTR(NIN, 0x15, "Natural Inertia", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
+	FTEC_TUNING_ATTR(FUL, 0x16, "FullForce", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
+
+enum ftec_tuning_attrs_enum {
+#define FTEC_TUNING_ATTR(id, addr, desc, conv_to, conv_from, min, max) \
+	id,
+FTEC_TUNING_ATTRS
+	FTEC_TUNING_ATTR_NONE
+#undef FTEC_TUNING_ATTR
 };
 
 #endif
