+++ hid-ftecff.c
@@ -43,100 +43,14 @@ static int profile = 1;
 module_param(profile, int, 0660);
 MODULE_PARM_DESC(profile, "Enable profile debug messages.");
 
-#define FTEC_TUNING_ATTRS \
-	FTEC_TUNING_ATTR(SLOT, 0x02, "Slot", ftec_conv_noop_to, ftec_conv_noop_from, 1, 5) \
-	FTEC_TUNING_ATTR(SEN, 0x03, "Sensivity", ftec_conv_sens_to, ftec_conv_sens_from, 90, 0) \
-	FTEC_TUNING_ATTR(FF, 0x04, "Force Feedback Strength", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
-	FTEC_TUNING_ATTR(SHO, 0x05, "Wheel Vibration Motor", ftec_conv_times_ten, ftec_conv_div_ten, 0, 100) \
-	FTEC_TUNING_ATTR(BLI, 0x06, "Break Level Indicator", ftec_conv_noop_to, ftec_conv_noop_from, 0, 101) \
-	FTEC_TUNING_ATTR(DRI, 0x09, "Drift Mode", ftec_conv_signed_to, ftec_conv_noop_from, -5, 3) \
-	FTEC_TUNING_ATTR(FOR, 0x0a, "Force Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
-	FTEC_TUNING_ATTR(SPR, 0x0b, "Spring Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
-	FTEC_TUNING_ATTR(DPR, 0x0c, "Damper Effect Strength", ftec_conv_times_ten, ftec_conv_div_ten, 0, 120) \
-	FTEC_TUNING_ATTR(NDP, 0x0d, "Natural Damber", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
-	FTEC_TUNING_ATTR(NFR, 0x0e, "Natural Friction", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
-	FTEC_TUNING_ATTR(FEI, 0x11, "Force Effect Intensity", ftec_conv_noop_to, ftec_conv_steps_ten, 0, 100) \
-	FTEC_TUNING_ATTR(INT, 0x14, "FFB Interpolation Filter", ftec_conv_noop_to, ftec_conv_noop_from, 0, 20) \
-	FTEC_TUNING_ATTR(NIN, 0x15, "Natural Inertia", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
-	FTEC_TUNING_ATTR(FUL, 0x16, "FullForce", ftec_conv_noop_to, ftec_conv_noop_from, 0, 100) \
-
-enum ftec_tuning_attrs_enum {
-#define FTEC_TUNING_ATTR(id, addr, desc, conv_to, conv_from, min, max) \
-	id,
-FTEC_TUNING_ATTRS
-	FTEC_TUNING_ATTR_NONE
-#undef FTEC_TUNING_ATTR
-};
-
-struct ftec_tuning_attr_t {
-	const char* name;
-	const enum ftec_tuning_attrs_enum id;
-	const u8 addr;
-	const char* description;
-	int (*conv_to)(struct ftec_drv_data *, u8);
-	u8 (*conv_from)(struct ftec_drv_data *, int);
-	const int min;
-	const int max;
-};
-
-int ftec_conv_sens_to(struct ftec_drv_data *drv_data, u8 val) {
-	if (drv_data->max_range <= 1090) {
-		return val * 10;
-	}
-	if (val < (u8)0x8a) {
-		return 1080 + 10 * (0x100 + val - 0xed);
-	} else if (val >= (u8)0xed) {
-		return 1080 + 10 * (val - 0xed);
-	}
-	return 90 + 10 * (val - 0x8a);
-};
-
-u8 ftec_conv_sens_from(struct ftec_drv_data *drv_data, int val) {
-	if (drv_data->max_range <= 1090) {
-		return val / 10;
-	}
-	if (val >= 1080) {
-		// overflow of u8 is expected behavior
-		return 0xed + ((val - 1080) / 10);
-	}
-	return 0x8a + (val - 90) / 10;
-};
-
-int ftec_conv_times_ten(struct ftec_drv_data *drv_data, u8 val) {
-	return val * 10;
-};
-
-u8 ftec_conv_div_ten(struct ftec_drv_data *drv_data, int val) {
-	return val / 10;
-};
-
-u8 ftec_conv_steps_ten(struct ftec_drv_data *drv_data, int val) {
-	return 10 * (val / 10);
-}
-
-int ftec_conv_signed_to(struct ftec_drv_data *drv_data, u8 val) {
-	return (s8)val;
-};
-
-int ftec_conv_noop_to(struct ftec_drv_data *drv_data, u8 val) {
-	return val;
-};
-
-u8 ftec_conv_noop_from(struct ftec_drv_data *drv_data, int val) {
-	return val;
-};
-
-
-static const struct ftec_tuning_attr_t ftec_tuning_attrs[] = {
-#define FTEC_TUNING_ATTR(id, addr, desc, conv_to, conv_from, min, max) \
-	{ #id, id, addr, desc, conv_to, conv_from, min, max },
-FTEC_TUNING_ATTRS
-#undef FTEC_TUNING_ATTR
-};
-
-static ssize_t _ftec_tuning_show(struct device *dev, enum ftec_tuning_attrs_enum id, char *buf);
-static ssize_t _ftec_tuning_store(struct device *dev, enum ftec_tuning_attrs_enum id,
-				 const char *buf, size_t count);
+int ftec_tuning_classdev_register(struct device *parent, 
+		struct ftec_tuning_classdev *ftec_tuning_cdev);
+void ftec_tuning_classdev_unregister(struct ftec_tuning_classdev *ftec_tuning_cdev);
+ssize_t _ftec_tuning_show(struct device *dev,
+		enum ftec_tuning_attrs_enum id, char *buf);
+ssize_t _ftec_tuning_store(struct device *dev,
+		enum ftec_tuning_attrs_enum id,
+		const char *buf, size_t count);
 
 static const signed short ftecff_wheel_effects[] = {
 	FF_CONSTANT,
@@ -346,7 +260,7 @@ static ssize_t ftec_range_show(struct device *dev, struct device_attribute *attr
 /* Set range to user specified value, call appropriate function
  * according to the type of the wheel */
 static ssize_t ftec_range_store(struct device *dev, struct device_attribute *attr,
-				 const char *buf, size_t count)
+				const char *buf, size_t count)
 {
 	struct hid_device *hid = to_hid_device(dev);
 	struct ftec_drv_data *drv_data = hid_get_drvdata(hid);
@@ -385,8 +299,7 @@ static DEVICE_ATTR(range, S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH, ftec_
 
 
 /* Export the current wheel id */
-static ssize_t ftec_wheel_show(struct device *dev, struct device_attribute *attr,
-				char *buf)
+static ssize_t ftec_wheel_show(struct device *dev, struct device_attribute *attr, char *buf)
 {
 	struct ftec_drv_data *drv_data = hid_get_drvdata(to_hid_device(dev));
 	return scnprintf(buf, PAGE_SIZE, "0x%02x\n", drv_data->wheel_id);
@@ -395,7 +308,7 @@ static DEVICE_ATTR(wheel_id, S_IRUSR | S_IRGRP | S_IROTH, ftec_wheel_show, NULL)
 
 
 static ssize_t ftec_set_display(struct device *dev, struct device_attribute *attr,
-				 const char *buf, size_t count)
+				const char *buf, size_t count)
 {
 	struct hid_device *hid = to_hid_device(dev);
 	struct ftec_drv_data *drv_data;
@@ -458,135 +371,6 @@ static ssize_t ftec_set_display(struct device *dev, struct device_attribute *att
 	return count;
 }
 static DEVICE_ATTR(display, S_IWUSR | S_IWGRP, NULL, ftec_set_display);
-
-static int ftec_tuning_write(struct hid_device *hid, int addr, int val) {
-	struct ftec_drv_data *drv_data = hid_get_drvdata(hid);
-	drv_data->ftec_tuning_data[0] = 0xff;
-	drv_data->ftec_tuning_data[1] = 0x03;
-	drv_data->ftec_tuning_data[2] = 0x00;
-	drv_data->ftec_tuning_data[addr+1] = val;
-	return hid_hw_output_report(hid, &drv_data->ftec_tuning_data[0], FTEC_TUNING_REPORT_SIZE);
-}
-
-static int ftec_tuning_select(struct hid_device *hid, int slot) {
-	u8 *buf = kcalloc(FTEC_TUNING_REPORT_SIZE, sizeof(u8), GFP_KERNEL);
-	int ret;
-    
-	buf[0] = 0xff;
-	buf[1] = 0x03;
-	buf[2] = 0x01;
-	buf[3] = slot&0xff;
-
-	ret = hid_hw_output_report(hid, buf, FTEC_TUNING_REPORT_SIZE);
-	kfree(buf);
-	return ret;
-}
-
-static enum ftec_tuning_attrs_enum ftec_tuning_get_id(struct device_attribute *attr) {
-	int idx = 0;
-	for (; idx < sizeof(ftec_tuning_attrs); idx++) {
-		if (strcmp(attr->attr.name, ftec_tuning_attrs[idx].name) == 0) {
-			return ftec_tuning_attrs[idx].id;
-		}
-	}
-	dbg_hid("Unknown attribute %s\n", attr->attr.name);
-	return FTEC_TUNING_ATTR_NONE;
-}
-
-static ssize_t ftec_tuning_show(struct device *dev, struct device_attribute *attr, char *buf)
-{
-	enum ftec_tuning_attrs_enum id = ftec_tuning_get_id(attr);
-	if (id == FTEC_TUNING_ATTR_NONE) {
-		return -EINVAL;
-	}
-
-	return _ftec_tuning_show(dev, id, buf);
-}
-
-static ssize_t _ftec_tuning_show(struct device *dev, enum ftec_tuning_attrs_enum id, char *buf)
-{
-	struct hid_device *hid = to_hid_device(dev);
-	struct ftec_drv_data *drv_data = hid_get_drvdata(hid);
-	const struct ftec_tuning_attr_t *tuning_attr = &ftec_tuning_attrs[id];
-	return scnprintf(buf, PAGE_SIZE, "%i\n", tuning_attr->conv_to(drv_data, drv_data->ftec_tuning_data[tuning_attr->addr+1]));
-}
-
-static ssize_t ftec_tuning_store(struct device *dev, struct device_attribute *attr,
-				 const char *buf, size_t count)
-{
-	enum ftec_tuning_attrs_enum id = ftec_tuning_get_id(attr);
-	if (id == FTEC_TUNING_ATTR_NONE) {
-		return -EINVAL;
-	}
-	return _ftec_tuning_store(dev, id, buf, count);
-}
-
-static ssize_t _ftec_tuning_store(struct device *dev, enum ftec_tuning_attrs_enum id,
-				 const char *buf, size_t count)
-{
-	struct hid_device *hid = to_hid_device(dev);
-	struct ftec_drv_data *drv_data = hid_get_drvdata(hid);
-	const struct ftec_tuning_attr_t *tuning_attr = &ftec_tuning_attrs[id];
-	int val, _max = tuning_attr->max;
-
-	if (kstrtos32(buf, 0, &val) != 0) {
-		hid_err(hid, "Invalid value %s!\n", buf);
-		return -EINVAL;
-	}
-
-	/* special case for SEN, max value is device specific */
-	if (id == SEN) {
-		_max = drv_data->max_range;
-		/* set max value if 0 is given */
-		if (val == 0) {
-			val = _max;
-		}
-	}
-
-	/* check if value is in range */
-	if (val < tuning_attr->min || val > _max) {
-		hid_err(hid, "Value %i out of range [%i, %i]!\n", val, tuning_attr->min, _max);
-		return -EINVAL;
-	}
-	
-	/* convert value to device specific value */
-	val = tuning_attr->conv_from(drv_data, val);
-	dbg_hid(" ... ftec_tuning_store %s %i\n", tuning_attr->name, val);
-
-	if (id == SLOT) {
-		if (ftec_tuning_select(hid, val) < 0) {
-			return -EIO;
-		}
-	} else {
-		if (ftec_tuning_write(hid, tuning_attr->addr, val) < 0) {
-			return -EIO;
-		}
-	}
-	return count;
-}
-
-static ssize_t ftec_tuning_reset(struct device *dev, struct device_attribute *attr,
-				 const char *buf, size_t count)
-{
-	struct hid_device *hid = to_hid_device(dev);
-	u8 *buffer = kcalloc(FTEC_TUNING_REPORT_SIZE, sizeof(u8), GFP_KERNEL);
-	int ret;
-    	
-	// request current values
-	buffer[0] = 0xff;
-	buffer[1] = 0x03;
-	buffer[2] = 0x04;
-
-	ret = hid_hw_output_report(hid, buffer, FTEC_TUNING_REPORT_SIZE);
-	
-	return count;
-}
-
-static DEVICE_ATTR(RESET, S_IWUSR  | S_IWGRP, NULL, ftec_tuning_reset);
-#define FTEC_TUNING_ATTR(id, addr, desc, conv_to, conv_from, min, max) \
-	static DEVICE_ATTR(id, S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH, ftec_tuning_show, ftec_tuning_store);
-FTEC_TUNING_ATTRS
-#undef FTEC_TUNING_ATTR
 
 #ifdef CONFIG_LEDS_CLASS
 static void ftec_set_leds(struct hid_device *hid, u16 leds)
@@ -1301,37 +1085,7 @@ int ftecff_init(struct hid_device *hdev) {
 	CREATE_SYSFS_FILE(wheel_id)
 
 	if (drv_data->quirks & FTEC_TUNING_MENU) {
-		CREATE_SYSFS_FILE(RESET)
-		CREATE_SYSFS_FILE(SLOT)
-		CREATE_SYSFS_FILE(SEN)
-		CREATE_SYSFS_FILE(FF)
-		CREATE_SYSFS_FILE(FEI)
-		CREATE_SYSFS_FILE(FOR)
-		CREATE_SYSFS_FILE(SPR)
-		CREATE_SYSFS_FILE(DPR)
-
-		if (hdev->product == CSL_ELITE_WHEELBASE_DEVICE_ID || 
-		    hdev->product == CSL_ELITE_PS4_WHEELBASE_DEVICE_ID) {
-			CREATE_SYSFS_FILE(DRI)
-		}
-		if (hdev->product == CSL_ELITE_WHEELBASE_DEVICE_ID || 
-		    hdev->product == CSL_ELITE_PS4_WHEELBASE_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD1_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD2_DEVICE_ID) {
-			CREATE_SYSFS_FILE(BLI)
-			CREATE_SYSFS_FILE(SHO)
-		}
-		if (hdev->product == CSL_DD_WHEELBASE_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD1_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD2_DEVICE_ID) {
-			CREATE_SYSFS_FILE(NDP)
-			CREATE_SYSFS_FILE(NFR)
-			CREATE_SYSFS_FILE(NIN)
-			CREATE_SYSFS_FILE(INT)
-		}	
-		if (hdev->product == CSL_DD_WHEELBASE_DEVICE_ID) {
-			CREATE_SYSFS_FILE(FUL)
-		}
+		ftec_tuning_classdev_register(&hdev->dev, &drv_data->tuning);
 	}
 
 #ifdef CONFIG_LEDS_CLASS
@@ -1366,37 +1120,7 @@ void ftecff_remove(struct hid_device *hdev)
 #define REMOVE_SYSFS_FILE(name) device_remove_file(&hdev->dev, &dev_attr_##name); \
 
 	if (drv_data->quirks & FTEC_TUNING_MENU) {
-		REMOVE_SYSFS_FILE(RESET)
-		REMOVE_SYSFS_FILE(SLOT)
-		REMOVE_SYSFS_FILE(SEN)
-		REMOVE_SYSFS_FILE(FF)
-		REMOVE_SYSFS_FILE(FEI)
-		REMOVE_SYSFS_FILE(FOR)
-		REMOVE_SYSFS_FILE(SPR)
-		REMOVE_SYSFS_FILE(DPR)
-
-		if (hdev->product == CSL_ELITE_WHEELBASE_DEVICE_ID || 
-		    hdev->product == CSL_ELITE_PS4_WHEELBASE_DEVICE_ID) {
-			REMOVE_SYSFS_FILE(DRI)
-		}
-		if (hdev->product == CSL_ELITE_WHEELBASE_DEVICE_ID || 
-		    hdev->product == CSL_ELITE_PS4_WHEELBASE_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD1_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD2_DEVICE_ID) {
-			REMOVE_SYSFS_FILE(BLI)
-			REMOVE_SYSFS_FILE(SHO)
-		}
-		if (hdev->product == CSL_DD_WHEELBASE_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD1_DEVICE_ID ||
-		    hdev->product == PODIUM_WHEELBASE_DD2_DEVICE_ID) {
-			REMOVE_SYSFS_FILE(NDP)
-			REMOVE_SYSFS_FILE(NFR)
-			REMOVE_SYSFS_FILE(NIN)
-			REMOVE_SYSFS_FILE(INT)
-		}	
-		if (hdev->product == CSL_DD_WHEELBASE_DEVICE_ID) {
-			REMOVE_SYSFS_FILE(FUL)
-		}
+		ftec_tuning_classdev_unregister(&drv_data->tuning);
 	}
 
 #ifdef CONFIG_LEDS_CLASS
@@ -1422,9 +1146,9 @@ int ftecff_raw_event(struct hid_device *hdev, struct hid_report *report, u8 *dat
 	struct ftec_drv_data *drv_data = hid_get_drvdata(hdev);
 	if (data[0] == 0xff && size == FTEC_TUNING_REPORT_SIZE) {
 		// shift by 1 so that we can use this as the buffer when writing back to the device
-		memcpy(&drv_data->ftec_tuning_data[0] + 1, data, sizeof(drv_data->ftec_tuning_data) - 1);
+		memcpy(&drv_data->tuning.ftec_tuning_data[0] + 1, data, sizeof(drv_data->tuning.ftec_tuning_data) - 1);
 		// notify userspace about value change
-		kobject_uevent(&hdev->dev.kobj, KOBJ_CHANGE);
+		kobject_uevent(&drv_data->tuning.dev->kobj, KOBJ_CHANGE);
 	} else if (data[0] == 0x01) {
 		// TODO: detect wheel change and react on it in some way?
 		bool changed = drv_data->wheel_id != data[0x1f];
